Make calendar relation optional on fetching event

+++ app.js
@@ -52,7 +52,7 @@ app.get('/calendar-events/:id/ms-event', async function(req, res, next) {
     const eventId = req.params.id;
     const event = await getCalendarEvent(eventId);
 
-    if (event && event['ms-identifier']) {
+    if (event && event['calendar'] && event['ms-identifier']) {
       const graphApi = new GraphApiClient(sessionUri);
       const msCalendarId = calendarManager.getMsCalendarId(event.calendar);
       const msEvent = await graphApi.getCalendarEvent(msCalendarId, event);
@@ -82,6 +82,9 @@ app.get('/calendar-events/:id/ms-event', async function(req, res, next) {
       if (!event) {
         console.log(`No calendar-event found with id ${eventId} in triplestore`);
         return res.status(404).send();
+      } else if (!event['calendar']) {
+        console.log(`Event with id ${eventId} is not linked to an agenda in the triplestore.`);
+        return res.status(424).send();
       } else {
         console.log(`Event with id ${eventId}, but without MS-event found in triplestore. Probably an Access-mastered event.`);
         return res.status(200).send({ data: null });
@@ -102,7 +105,11 @@ app.patch('/calendar-events/:id', async function(req, res, next) {
     const eventId = req.params.id;
     const event = await getCalendarEvent(eventId);
 
-    if (event && event['ms-identifier']) {
+    if (!event['calendar']) {
+      event['calendar'] = await calendarManager.determineCalendar(event);
+    }
+
+    if (event && event['calendar'] && event['ms-identifier']) {
       const graphApi = new GraphApiClient(sessionUri);
       const payload = req.body.data.attributes;
       payload.id = eventId;
@@ -128,6 +135,9 @@ app.patch('/calendar-events/:id', async function(req, res, next) {
       if (!event) {
         console.log(`No calendar-event found with id ${eventId} in triplestore`);
         return res.status(404).send();
+      } else if (!event['calendar']) {
+        console.log(`Event with id ${eventId} is not linked to an agenda in the triplestore.`);
+        return res.status(424).send();
       } else { // event without ms-identifier found in triplestore
         console.log(`Cannot update Access-mastered event.`);
         return res.status(409).send();
@@ -151,7 +161,7 @@ app.delete('/calendar-events/:id', async function(req, res, next) {
       // delete in triplestore first such that delta's can already
       // be processed by other service (e.g. mu-cl-resources)
       await deleteCalendarEvent(event.uri);
-      if (event['ms-identifier']) {
+      if (event['calendar'] && event['ms-identifier']) {
         const graphApi = new GraphApiClient(sessionUri);
         const msCalendarId = calendarManager.getMsCalendarId(event.calendar);
         await graphApi.deleteCalendarEvent(msCalendarId, event['ms-identifier']);
